Moved remaining glue.clients.layer_artist to glue.core.layer_artist and rename LayerArtist to MatplotlibLayerArtist
