+++ glue/clients/__init__.py
@@ -1,3 +1,1 @@
 from __future__ import absolute_import, division, print_function
-
-from .layer_artist import LayerArtist
