+++ glue/viewers/common/viz_client.py
@@ -4,7 +4,7 @@ import matplotlib.pyplot as plt
 
 from glue.core import Data
 from glue.core.client import Client
-from glue.clients.layer_artist import LayerArtistContainer
+from glue.core.layer_artist import LayerArtistContainer
 from glue.utils.matplotlib import freeze_margins
